client_win32/requirements.txt
# HTTP 客户端
tornado>=6.0

# 音频
sounddevice>=0.4.6
numpy>=1.23,<2.0

# 热键监听
keyboard>=0.13.5

# 键盘输入
pynput>=1.7.6

# Windows API
pywin32>=305

# 系统托盘
pystray>=0.19.0
Pillow>=10.0.0

# 配置
PyYAML>=6.0

# 打包
pyinstaller>=6.0